add list cases
