+++ src/use-case/case/list-cases.use-case.js
@@ -1,0 +1,7 @@
+import { caseRepository } from "../../repository/case.repository";
+import { CaseModel } from "../../model/case.model";
+
+export const listCasesUseCase = async (listQuery) => {
+  const cases = await caseRepository.findCases(listQuery);
+  return cases.data.map((caseItem) => CaseModel.existingCase(caseItem));
+};
